+++ ParaView/ParaView/vtkCTHFractal.h
@@ -71,11 +71,11 @@ protected:
                 int x0, int y0, int z0);
 
   int LineTest2(float x0, float y0, float z0, 
-               float x1, float y1, float z1,
-               float bds[6]); 
+                float x1, float y1, float z1,
+                double bds[6]); 
   int LineTest(float x0, float y0, float z0, 
                float x1, float y1, float z1,
-               float bds[6], int level, int target); 
+               double bds[6], int level, int target); 
 
   void SetBlockInfo(int blockId, int level,
                     int x0, int y0, int z0);
